requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.2
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-bootstrap4==23.1
django-bootstrap5==23.1
django-crispy-forms==2.0
django-sendgrid-v5==1.2.2
djangorestframework==3.14.0
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
mysqlclient==2.1.1
oauthlib==3.2.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2023.3
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
sendgrid==6.10.0
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.1
soupsieve==2.4
sqlparse==0.4.3
starkbank-ecdsa==2.2.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
